🎨 Refactor matrix eigendecomposition methods
