+++ mlir/lib/Dialect/QCO/Utils/CMakeLists.txt
@@ -6,29 +6,15 @@
 #
 # Licensed under the MIT License
 
-add_mlir_library(
-  MLIRQCOMatrix
-  PARTIAL_SOURCES_INTENDED
-  Matrix.cpp
-  Eigensolver.cpp
-  LINK_LIBS
-  PUBLIC
-  MLIRSupport)
+add_mlir_library(MLIRQCOMatrix PARTIAL_SOURCES_INTENDED Matrix.cpp LINK_LIBS PUBLIC MLIRSupport)
 
 mqt_mlir_target_use_project_options(MLIRQCOMatrix)
 
-target_sources(
-  MLIRQCOMatrix
-  PUBLIC FILE_SET
-         HEADERS
-         BASE_DIRS
-         ${MQT_MLIR_SOURCE_INCLUDE_DIR}
-         FILES
-         ${MQT_MLIR_SOURCE_INCLUDE_DIR}/mlir/Dialect/QCO/Utils/Matrix.h
-         ${MQT_MLIR_SOURCE_INCLUDE_DIR}/mlir/Dialect/QCO/Utils/Eigensolver.h)
+target_sources(MLIRQCOMatrix PUBLIC FILE_SET HEADERS BASE_DIRS ${MQT_MLIR_SOURCE_INCLUDE_DIR} FILES
+                                    ${MQT_MLIR_SOURCE_INCLUDE_DIR}/mlir/Dialect/QCO/Utils/Matrix.h)
 
 file(GLOB_RECURSE UTILS_CPP "${CMAKE_CURRENT_SOURCE_DIR}/*.cpp")
-list(FILTER UTILS_CPP EXCLUDE REGEX "(Matrix|Eigensolver)\\.cpp$")
+list(FILTER UTILS_CPP EXCLUDE REGEX "Matrix\\.cpp$")
 
 add_mlir_dialect_library(
   MLIRQCOUtils
@@ -48,7 +34,7 @@ mqt_mlir_target_use_project_options(MLIRQCOUtils)
 
 # collect header files
 file(GLOB_RECURSE UTILS_HEADERS_SOURCE "${MQT_MLIR_SOURCE_INCLUDE_DIR}/mlir/Dialect/QCO/Utils/*.h")
-list(FILTER UTILS_HEADERS_SOURCE EXCLUDE REGEX "(Matrix|Eigensolver)\\.h$")
+list(FILTER UTILS_HEADERS_SOURCE EXCLUDE REGEX "Matrix\\.h$")
 file(GLOB_RECURSE UTILS_HEADERS_BUILD "${MQT_MLIR_BUILD_INCLUDE_DIR}/mlir/Dialect/QCO/Utils/*.inc")
 
 # add public headers using file sets
